feat(video): creating the video shared block
