+++ express/blocks/tutorials/tutorials.css
@@ -74,7 +74,7 @@ main .tutorials h3 {
   font-weight: 700;
 }
 
-main .tutorials-overlay {
+main .video-overlay {
   position: fixed;
   top: 0;
   bottom: 0;
@@ -87,20 +87,20 @@ main .tutorials-overlay {
   z-index: 10;
 }
 
-main .tutorials-overlay .tutorials-overlay-video {
+main .video-overlay .video-overlay-video {
   width: 90vw;
   height: calc(90vw * 9 / 16);
   max-height: 100vh;
 }
 
-main .tutorials-overlay .tutorials-overlay-video.vimeo {
+main .video-overlay .video-overlay-video.vimeo {
   width: 90vh;
   max-width: 90vw;
   height: 90vw;
   max-height: 90vh;
 }
 
-main .tutorials-overlay .tutorials-overlay-video iframe {
+main .video-overlay .video-overlay-video iframe {
   width: 100%;
   height: 100%;
 }
